+++ Quick_Transfer/Plugin.cs
@@ -34,7 +34,7 @@ public class Plugin : BaseUnityPlugin
 {
     private const string PluginGuid = "crispywhips.quick_transfer";
     public const string PluginName = "Quick_Transfer";
-    public const string PluginVersion = "1.8.0";
+    public const string PluginVersion = "1.8.1";
 
     internal new static ManualLogSource Logger;
     private static Harmony _harmony;